test: fix EmptyPlaceholder.test.jsx and convert to typescript
